move files to right places
