+++ tests/integration/snapshots/snapshot-configuration.test.js
