+++ tests/unit/snapshot/InitialSnapshotManager.test.js
