+++ tests/unit/snapshot/MemorySnapshotStore.test.js
