+++ tests/unit/snapshot/SnapshotConfigManager.test.js
