+++ tests/unit/snapshot/SnapshotManager.test.js
