+++ tests/unit/snapshot/SnapshotTrigger.test.js
