+++ tests/unit/snapshot/ToolManagerIntegration.test.js
