+++ tests/unit/snapshot/ToolMonitor.test.js
